This script extracts images of scientists.
The script first edits the html and adds a div to separate list of scientists from other data and saves it in a "out.html". "out.html" is used again to iterate over all links and hence downloads all images and generates a valid json file. The now not required "out.html" is deleted at the end.